move rekey to jsonutil from parallel.util
also make rekey accept basestring, not just str.

closes gh-532
